Fix [ 3402314 ] non-ASCII whitespace and punctuation around inline markup.
Revision of the rules for allowed characters around inline markup
start-string and end-string: Keep the carefully crafted ASCII-character set
but add Unicode categories to the sets of allowed characters. This keeps
the number of "false positives" requiring escaping low while making the rules
simpler and international.

This is a feature change.

git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@7243 929543f6-e4f2-0310-98a6-ba3bd3dd1d04