New setting `stylesheet_dirs`.
Comma-separated list of directories where stylesheets are found. Used by
`stylesheet_path` when expanding relative path arguments.

Now, it is easy to add a custom stylesheet to Docutils default
stylesheet with, e.g., --stylesheet_path="html4css1.css, mystyle.css"
    
Changed behaviour of the default settings: 
if there is a file "html4css1.css" in the working directory of the
process at launch, it is used instead of the one provided by Docutils
in the writer source directory.


git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@7618 929543f6-e4f2-0310-98a6-ba3bd3dd1d04